pyproject.toml
[project]
name = "django-aiogram"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Suhrob", email = "mlkv.dev@gmail.com" }
]
dependencies = [
    "django>=5.2",
    "pillow==10.3.0",
    "psycopg2==2.9.9; os_name == 'linux'",
    "psycopg2-binary==2.9.9; os_name != 'linux'",
    "aiogram>=3.19.0",
    "ipython>=9.0.2",
    "python-decouple>=3.8",
    "loguru>=0.7.3",
    "phonenumbers>=9.0.2",
    "redis>=5.2.1",
    "django-redis>=5.4.0",
    "django-phonenumber-field[phonenumbers]>=8.0.0",
    "uvicorn>=0.34.0",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "apscheduler>=3.11.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
